requirements.txt
cachecontrol==0.12.5
Flask==1.1.1
google-api-python-client==1.7.11
google-cloud-datastore==1.10.0
jsonschema==3.1.1
requests==2.22.0
